+++ CogPar/src/main/java/co/uk/cogitolearning/cogpar/Algorithms.java
@@ -6,7 +6,7 @@ import java.util.Iterator;
 
 public class Algorithms {
     static void setVariable(ExpressionNode root, String name, double value) {
-        SetVariable piVisitor = new SetVariable(name, value);
+        VariableSetterVisitor piVisitor = new VariableSetterVisitor(name, value);
         Iterator<ExpressionNode> it = root.iterator();
         while (it.hasNext())
             it.next().accept(piVisitor);
