+++ CogPar/src/main/java/co/uk/cogitolearning/cogpar/Calculator.java
@@ -24,41 +24,15 @@
 
 package co.uk.cogitolearning.cogpar;
 
-import co.uk.cogitolearning.cogpar.parser.Parser;
 import co.uk.cogitolearning.cogpar.tree.ExpressionNode;
+import lombok.experimental.UtilityClass;
 
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.Stack;
 
-/**
- * Test the Parser
- */
-public class Calculator {
-
-    /**
-     * The main method to test the functionality of the parser
-     */
-    public static void main(String[] args) {
-
-        String exprstr = "2*(1+sin(0.5 * pi))^2"; //=8  with pi/2 is problem because DivExpressionNode was introduced. Parser has to be fixed.
-//    String exprstr = "2*3";
-        if (args.length > 0) exprstr = args[0];
-
-        Parser parser = new Parser();
-        try {
-            ExpressionNode expr = parser.parse(exprstr);
-            Algorithms.setVariable(expr, "pi", Math.PI);
-
-            System.out.println("The value of the expression is " + calculate(expr));
-
-        } catch (ParserException e) {
-            System.out.println(e.getMessage());
-        } catch (EvaluationException e) {
-            System.out.println(e.getMessage());
-        }
-    }
-
+@UtilityClass
+class Calculator {
 
     static double calculate(ExpressionNode expr) {
         ArrayList<ExpressionNode> polishNotationList = new ArrayList<>();
